+++ internal/api/auth/authorize.go
@@ -147,7 +147,8 @@ func (m *Module) AuthorizeGETHandler(c *gin.Context) {
 	// the authorize template will display a form to the user where they can get some information
 	// about the app that's trying to authorize, and the scope of the request.
 	// They can then approve it if it looks OK to them, which will POST to the AuthorizePOSTHandler
-	c.HTML(http.StatusOK, "authorize.tmpl", gin.H{
+	c.HTML(http.StatusOK, "page.tmpl", gin.H{
+		"template":   "authorize.tmpl",
 		"appname":    app.Name,
 		"appwebsite": app.Website,
 		"redirect":   redirect,
