+++ testrig/router.go
@@ -59,6 +59,5 @@ func NewTestRouter(db db.DB) *router.Router {
 
 // ConfigureTemplatesWithGin will panic on any errors related to template loading during tests
 func ConfigureTemplatesWithGin(engine *gin.Engine, templatePath string) {
-	router.LoadTemplateFunctions(engine)
 	engine.LoadHTMLGlob(filepath.Join(templatePath, "*"))
 }
