+++ web/template/404.tmpl
@@ -17,7 +17,6 @@
 // along with this program.  If not, see <http://www.gnu.org/licenses/>.
 */ -}}
 
-{{ template "header.tmpl" .}}
 <main>
     <section>
         <h1>404: Page Not Found</h1>
@@ -35,5 +34,3 @@
         </p>
     </section>
 </main>
-
-{{ template "footer.tmpl" .}}
