--- web/template/footer.tmpl
@@ -1,46 +1,0 @@
-{{- /*
-// GoToSocial
-// Copyright (C) GoToSocial Authors admin@gotosocial.org
-// SPDX-License-Identifier: AGPL-3.0-or-later
-//
-// This program is free software: you can redistribute it and/or modify
-// it under the terms of the GNU Affero General Public License as published by
-// the Free Software Foundation, either version 3 of the License, or
-// (at your option) any later version.
-//
-// This program is distributed in the hope that it will be useful,
-// but WITHOUT ANY WARRANTY; without even the implied warranty of
-// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-// GNU Affero General Public License for more details.
-//
-// You should have received a copy of the GNU Affero General Public License
-// along with this program.  If not, see <http://www.gnu.org/licenses/>.
-*/ -}}
-
-        <!-- footer.tmpl -->
-        </div>
-        <footer>
-            <div id="version">
-                <a name="Source code" href="https://github.com/superseriousbusiness/gotosocial">
-                    GoToSocial <span class="accent">{{.instance.Version}}</span>
-                </a>
-            </div>
-            {{ if .instance.ContactAccount }} 
-                <div id="contact">
-                    Contact: <a href="{{.instance.ContactAccount.URL}}" class="nounderline">{{.instance.ContactAccount.Username}}</a><br>
-                </div>
-            {{ end }}
-            {{ if .instance.Email }} 
-                <div id="email">
-                    Email: <a href="mailto:{{.instance.Email}}" class="nounderline">{{.instance.Email}}</a><br>
-                </div>
-            {{ end }}
-        </footer>
-    </div>
-    {{ if .javascript }}
-    {{ range .javascript }}
-        <script src="{{.}}"></script>
-    {{ end }}
-    {{ end }}
-</body>
-</html>
