+++ web/template/profile.tmpl
@@ -17,75 +17,72 @@
 // along with this program.  If not, see <http://www.gnu.org/licenses/>.
 */ -}}
 
-{{ template "header.tmpl" .}}
-
+{{- with . }}
 <main class="profile">
     <div class="header">
         <div class="header-image">
-            {{ if .account.Header }}
-            <img src="{{.account.Header}}" alt="" />
-            {{ end }}
+            {{- if .account.Header }}
+            <img src="{{- .account.Header -}}" alt="" />
+            {{- end }}
         </div>
         <div class="basic-info" aria-hidden="true">
-            <a class="avatar" href="{{.account.Avatar}}">
-                <img src="{{.account.Avatar}}" alt="">
+            <a class="avatar" href="{{- .account.Avatar -}}">
+                <img src="{{- .account.Avatar -}}" alt="">
             </a>
             <span class="displayname text-cutoff">
-                {{if .account.DisplayName}}
-                {{emojify .account.Emojis (escape .account.DisplayName)}}
-                {{else}}
-                {{.account.Username}}
-                {{end}}
+                {{- if .account.DisplayName -}}
+                {{ emojify .account.Emojis (escape .account.DisplayName) }}
+                {{- else }}
+                {{ .account.Username }}
+                {{- end }}
             </span>
             <span class="username text-cutoff">@{{.account.Username}}@{{.instance.AccountDomain}}</span>
             {{- /* Only render account role if 1. it's present and 2. it's not equal to the standard 'user' role */ -}}
-            {{ if and (.account.Role) (ne .account.Role.Name "user") }}
-            <div class="role {{ .account.Role.Name }}">
+            {{- if and (.account.Role) (ne .account.Role.Name "user") }}
+            <div class="role {{ .account.Role.Name -}}">
                 {{ .account.Role.Name }}
             </div>
-            {{ end }}
+            {{- end }}
         </div>
         <div class="sr-only">
             Profile for
-            {{if .account.DisplayName}}{{.account.DisplayName}}{{else}}{{.account.Username}}{{end}}.
-            Username @{{.account.Username}}, {{.instance.AccountDomain}}.
-            {{ if and (.account.Role) (ne .account.Role.Name "user") }}
+            {{- if .account.DisplayName }}
+            {{ .account.DisplayName }}
+            {{- else }}
+            {{ .account.Username }}
+            {{- end }}.
+            Username @{{- .account.Username -}}, {{ .instance.AccountDomain -}}.
+            {{- if and (.account.Role) (ne .account.Role.Name "user") }}
             Role: {{ .account.Role.Name }}
-            {{ end }}
+            {{- end }}
         </div>
     </div>
-
     <div class="column-split">
-
         <section class="about-user">
             <div class="col-header">
                 <h1>About</h1>
             </div>
-
             <div class="fields">
-                {{ range .account.Fields }}
+                {{- range .account.Fields }}
                 <div class="field">
-                    <b>{{emojify $.account.Emojis (noescape .Name)}}</b>
-                    <span>{{emojify $.account.Emojis (noescape .Value)}}</span>
+                    <b>{{- emojify $.account.Emojis (noescape .Name) -}}</b>
+                    <span>{{- emojify $.account.Emojis (noescape .Value) -}}</span>
                 </div>
-                {{ end }}
+                {{- end }}
             </div>
-
             <div class="bio">
-                {{ if .account.Note }}
-                {{emojify .account.Emojis (noescape .account.Note)}}
-                {{else}}
+                {{- if .account.Note }}
+                {{ emojify .account.Emojis (noescape .account.Note) }}
+                {{- else }}
                 This GoToSocial user hasn't written a bio yet!
-                {{end}}
+                {{- end }}
             </div>
-
             <div class="sr-only" role="group">
                 <span>Joined on {{.account.CreatedAt | timestampVague}}.</span>
                 <span>{{.account.StatusesCount}} post{{if .account.StatusesCount | eq 1 | not}}s{{end}}.</span>
                 <span>Followed by {{.account.FollowersCount}}.</span>
                 <span>Following {{.account.FollowingCount}}.</span>
             </div>
-
             <div class="accountstats" aria-hidden="true">
                 <b>Joined</b><time datetime="{{.account.CreatedAt}}">{{.account.CreatedAt | timestampVague}}</time>
                 <b>Posts</b><span>{{.account.StatusesCount}}</span>
@@ -93,53 +90,48 @@
                 <b>Following</b><span>{{.account.FollowingCount}}</span>
             </div>
         </section>
-
         <section class="toots">
-            {{ if .pinned_statuses }}
+            {{- if .pinned_statuses }}
             <div class="col-header">
                 <h2>Pinned posts</h2>
                 <a href="#recent">jump to recent</a>
             </div>
             <section class="thread">
-                {{ range .pinned_statuses }}
-                <article class="toot expanded" id="{{.ID}}">
-                    {{ template "status.tmpl" .}}
+                {{- range .pinned_statuses }}
+                <article class="toot expanded" id="{{- .ID -}}">
+                    {{ include "status.tmpl" . | indent 5 | noescape }}
                 </article>
-                {{ end }}
+                {{- end }}
             </section>
-            {{ end }}
-
+            {{- end }}
             <div class="col-header">
                 <h2 id="recent" tabindex="-1">Recent posts</h2>
-                {{ if .rssFeed }}
-                <a href="{{ .rssFeed }}" class="rss-icon" aria-label="RSS feed">
+                {{- if .rssFeed }}
+                <a href="{{- .rssFeed -}}" class="rss-icon" aria-label="RSS feed">
                     <i class="fa fa-rss-square" aria-hidden="true"></i>
                 </a>
-                {{ end }}
+                {{- end }}
             </div>
-
             <section class="thread">
-                {{ if not .statuses }}
+                {{- if not .statuses }}
                 <div data-nosnippet class="nothinghere">Nothing here!</div>
-                {{ else }}
-                {{ range .statuses }}
-                <article class="toot expanded" id="{{.ID}}">
-                    {{ template "status.tmpl" .}}
+                {{- else }}
+                {{- range .statuses }}
+                <article class="toot expanded" id="{{- .ID -}}">
+                    {{ include "status.tmpl" . | indent 5 | noescape }}
                 </article>
-                {{ end }}
-                {{ end }}
+                {{- end }}
+                {{- end }}
             </section>
-
             <div class="backnextlinks">
-                {{ if .show_back_to_top }}
-                <a href="/@{{ .account.Username }}">Back to top</a>
-                {{ end }}
-                {{ if .statuses_next }}
-                <a href="{{ .statuses_next }}" class="next">Show older</a>
-                {{ end }}
+                {{- if .show_back_to_top }}
+                <a href="/@{{- .account.Username -}}">Back to top</a>
+                {{- end }}
+                {{- if .statuses_next }}
+                <a href="{{- .statuses_next -}}" class="next">Show older</a>
+                {{- end }}
             </div>
         </section>
     </div>
 </main>
-
-{{ template "footer.tmpl" .}}
+{{- end }}
