+++ web/template/tag.tmpl
@@ -17,11 +17,9 @@
 // along with this program.  If not, see <http://www.gnu.org/licenses/>.
 */ -}}
 
-{{ template "header.tmpl" .}}
-
+{{- with . }}
 <main class="thread">
-    <h2 id="tag-name" tabindex="-1">#{{.tagName}}</h2>
+    <h2 id="tag-name" tabindex="-1">#{{- .tagName -}}</h2>
     <p>There's nothing here yet!</p>
 </main>
-
-{{ template "footer.tmpl" .}}
+{{- end }}
